Fix minPointer condition checking on Android's PanHandler
We've been checking minPointer constraint on bot POINTER_UP and POINTER_DOWN events. This wasn't right as when placing the fingers down the fingers will never touch the screen at exactly the same time and therefore before min fingers could be met there is an intermediate state in which we have fewer than min number of fingers. In that case MIN_POINTERS check would fail and so the handler. The change is to only be checking min_pointers if the handler already activated. Until it's active we can still expect user to place more fingers on the screen for the min pointers criteria to be met.

This fixes #99