Bottom sheet example and TapGestureHandler improvements (#143)
## Motivation
https://github.com/brentvatne/bottom-sheet-gesture-handler-experiment
Got inspired by @brentvatne code and made it possibile to implement bottom sheet in a bit hacky way using very long TapHandler. The problem was to to trigger scrolling scrollview at the end of extending a main card and not to trigger it when card is not opened fully.
Also then it became extremely easy to resolve another issue: https://github.com/kmagiera/react-native-gesture-handler/issues/146

## Changes
Initially I was trying to made it possible to track route (and maxDeltas) using 'normal' TapGestureHandler. Then I decided to implement new `TapGestureHandler` as quite unobvious extension of `UIPanGestureRecognizer` which gives us many ways of parameterization.
However, I wasn't really into extending out `PanGestureHandler` which seems to be unnecessary.
It gives a possibility to fail gesture whet its position pass `maxDeltaY` and then it allows scroll gesture to be triggered then. If it won't fail, scroll won't be scrolled.