Add support for references in React 16.03 (#157)
Motivation

Using string IDs for linking handlers has been a huge pain. The reason is that the IDs has to be globally unique to make your components reusable. This change makes it possible to replace string identifiers with a new mechanism for creating refs introduced in React 16.3 (https://reactjs.org/docs/refs-and-the-dom.html)

Changes

When processing waitFor or simultHandlers lists we try to figure out if the provided object is a React ref (by testing if .current field is set). If so then we extract handler tag directly from it.