Add Android logic for support multihandlers (#237)
## Motivation
Behaviour on Android and iOS in matter of handling two handler in the same time is completely different and calls for changes on Android site which are done in this commit
## Changes
As every gesture is being recognised by one big native handler and then pass to our handle we have to make decision to which handler each motion event (only part of motion event connected with current handler) should be passed.
As it wasn't possible to "extend" MotionEvent I need to make a proper adapter which is responsible to behave like a motion event for each recogniser. 
`GestureHandlerMotionEventAdapter` gives information about pointers connected with current handler. What's more it "blocks" events (by result of `updateMotionEventBeforeHandling`) which are not related to current handler.
Adapter needs also to pretend delivering `POINTER_DOWN` or `POINTER_UP` events to behave naturally. Code seems to be self-descriptive https://github.com/kmagiera/react-native-gesture-handler/pull/237/files#diff-cb2803f12c229320bf636e3d1371f370R31
Also logic of "cancelling" had to be changed as with two handlers "began" it was't possible to make "active" both of them. To make it work I allow to cancel only these "began" handlers which are connected with the same pointer https://github.com/kmagiera/react-native-gesture-handler/pull/237/files#diff-42183bb4b21fce222dc0e2baa2b0a1d3R493
As it wasn't to possible to simply override "MotionEvent" because of strange "final" methods it has made `scaleDetector` broken because of its requirement to handle "MotionEvent" type. So I made a custom scaleDetector based on native one.
 I need to make changes in `RotationDetector` which I suppose has crucial error https://github.com/kmagiera/react-native-gesture-handler/pull/237/files#diff-ddd7b851d29b2d899714ea8df12294c6R40 as `getX` in MotionEvent needs index of pointer (not id).
Another changes is connected with `VelocityTracker` as its method refers to whole event 
https://github.com/kmagiera/react-native-gesture-handler/pull/237/files#diff-cb2803f12c229320bf636e3d1371f370R102
Also added two examples which seems to be workable