Prevent interpolate when values are out of order (#301)
This is fix for broken behavior of Swipeable on newer version of React Native (where interpolate do not accept inputRange elements that are not growing/descending).

When `rigthWidth` is set to `0` (like when we load data to `Swipeable`) app is freezing for few seconds. 
I expect that similar behavior could be reproduced for `leftWidth`.