Force Touch  GH (#374)
Motivation
Add possibility of handling force touch gesture.

Changes
Add proper handler on iOS and no-op fallback on Android. On iOS for calculating force touch only first finger is concerned which is normal behaviour I observed on iOS. Force is expressed as value greater of equal 0 and lesser or equal 1.
I added two options of customisation.

minForce
minForce is a minimal pressure which has to be applied in order to activate handler.

maxForce
minForce is a maximal force which could be applied. If greater force applied, handler fails.

On releasing first finger gesture fails or ends (if has been activated already).

And added simple check method to determine is it possible to use forceTouch on current device which exporting RN's constant.