Copy classes methods properly in _refHandler (#398)
I noticed some methods missing from the gesture handler wrapped components since updating RN to latest master. It happens because a lot of core components were changed to use classes instead of React.createClass and the logic to copy methods did not handle classes properly.

This fixes it by looping through the prototype and using Object. getOwnPropertyNames to get methods instead of for ... in. This way it handles classes and regular objects properly.