Add touchables (#372)
Motivation
This PR provides a gesture-handler's native button backed implementation of Touchabnles. The goal is to provide a compatibility layer for Touchable components from RN core but make it backed by the gesture-handler components that integrates better with the native touch system.

Changes
In order to handle it I prepared following polyfills inspired by RN's core implementation. Using LongPressGestureHandler and TapGestureHandler and few new features and fixes in RNGH I managed to imitate RN's touchables.

How it works?
My implementation of Touchables is a states' machine which appears to be be a bit simpler than RN's implementation and has beed documented with comments in code.

Add examples
Added example for many use cases of Touchables (compared RN's and GN's touchables).

Differences
TouchableWithNativeFeedbacks are a bit different here and in RN's core. It's visible on long pressing. I decided not to change it since it would call for a lot of fixes in native node. However my implementation is closer do native behaviour. RN's implementation was manually triggering ripple effect via bridge from JS.