Handle case of null-parent (#385)
Motivation
This PR is a solution for rarely observed problem of parent-null issue on Android devices. If occurs when we want to extract position od handler when one of its parent has been detached.

Why it helps?
I believe that this issue might be related to some race-condition on Android devices when it comes to rendering some complicated structures. I could believe that event sending from componentWillUnmount might update GH after gesture event appear in some edge cases.

I cannot prove that it's actually a problem and it's actually solution, but it's not breaking and if it could make our users' apps more reliable, I'd like to have it merged.