Exclude the "Example" folder in the podspec
This needs to be done because the `s.source_files  = "ios/**/*.{h,m}"` was also picking up files inside `Example/ios/*`. Inside this folder, there is a file called `AppDelegate` which was causing duplicate symbol errors when using this library through cocoapods.

This is fixed by explicitly excluding the entire `Example` folder.