Remove peerDependencies for web (#487)
react-native-web shouldn't be a peerDep, and I'm not sure why react-art is there. I think we could also remove react and react-native and just put them as `'*'` and specify more clear version requirements in the README / installation instructions, because it all just gets lots in the peerDependencies warning mess that react-native is in.