[Android][Pinch] Cache focalX/focalY values and only update when two fingers present
Fixes #546. Android's [ScaleGestureDetector](https://developer.android.com/reference/android/view/ScaleGestureDetector.html#getFocusX()) says "the focal point is between each of the pointers forming the gesture". However, that means that at the end of the gesture, when the first finger is removed from the screen, the focal point will suddenly be redefined as the location of the second finger, rather than being between the two fingers.

This PR prevents the focal point from being updated when only one finger is present.