Replace propType dependency with string literals (#573)
I looked through the files and tried to replace what propType dependencies I could find with string literals. I also tried out the Example app with babel-plugin-transform-react-remove-prop-types and it seemed to me like it was working (also verified that the propTypes actually was removed by printing them to the console and making sure they were undefined).

However, if someone with more experience with the code base could take a gander at the changes and verify that it is working that would be great, I'd if I missed anything and it accidentally got merged. Since I have limited experience with all the components I might have missed some regressions but I hope not.

Thanks in advance 🙂

This should fix #548.
Co-Authored-By: Michał Osadnik <micosa97@gmail.com>