Don't call deprecated `[RCTRootView cancelTouches]`
According to the message in the commit it should not be needed anymore https://github.com/facebook/react-native/commit/36307d87e1974aff1abac598da2fd11c4e8e23c1. Not sure if we have a good test case for this to make sure it doesn't break but I've used it in my app for a while and everything seems fine (although I don't really mix RN and RNGH gesture components).

I kept `didActivateInRootView` empty since it is needed by the protocol but it might be possible to remove it from the protocol also.