Fix #320: Unsupported event crash. (#845)
After recent changes in RN core we started seeing the well known "Unsupported event" crash.

The crash is due to the fact that gesture handler requires for all view types to accept geature-handler events, namely, onGestureHandlerEvent and onGestureHandlerStateChange. This is necessary for us to be able to deliver those events to RN Views and other components that need to respond to gestures. As RN core does not expose a method that'd allow us to register custom events for RCTView, we've been relying on a few workarounds that unfortunately only worked on some version of RN. The latest workaround was to override `genericDirectEventTypes` object from `UIManager` configuration. This object is then used during view registration to populate event config maps. We've been extending `UIManager.genericDirectEventTypes` at first until it got moved to `UImanager.getConstants().genericDirectEventTypes` -- after that change we started overriding `getConstants` method to return the extended object. For a reason I haven't had time to figure out this stopped working in RN 62 and the current workaround is to just extend the object returned from `getConstants()` relying on the assumption that this method always returns a reference to the same method.

As it was the case for the previous workaround this one still requires that we load gesture-handler module before *any* view registration is done. Once the views are registered it is no longer possible to update their event maps for their view configuration.