Export a component for gesture handler root view
Currently the user always needs to add code to wrap the root view in their MainActivity.java on Android, or in case of libraries such as wix's native navigation, wrap each screen in `gestureHandlerRootHOC`.

This works well for apps, but for libraries which use gesture handler, it often increases the setup steps for the user. In case of React Navigation and other libraries I work on. we have found that people always miss this configuration step and open issues about their gestures not working. This wrapper will let library authors avoid this extra step for the user, and will only need auto-linking to work.

- [ ] Actually test the code
- [ ] We can take care of nested views with context in JS, but what if user added it in both `MainActivity.java` and here? Should we remove the `Mainactivity.java` step and only use the component/hoc? Arguably it's easier setup than `MainActivity.java`.