Export a component for gesture handler root view (#925)
Currently, the user always needs to add code to wrap the root view in their MainActivity.java on Android, or in case of libraries such as Wix's native navigation, wrap each screen in gestureHandlerRootHOC.

This works well for apps, but for libraries that use gesture handler, it often increases the setup steps for the user. In the case of React Navigation and other libraries I work on, we have found that people always miss this configuration step and open issues about their gestures not working. This wrapper will let library authors avoid this extra step for the user by wrapping their library code in GestureHandlerRootView, and will only need auto-linking to work.

Overview

New GestureHandlerRootView component for libraries to use, takes care of nesting multiple root views with context API so only 1 of them is rendered and rest are replaced with View
Change gestureHandlerRootHOC to use the GestureHandlerRootView component which also fixes a platform inconsistency
Fix the native RNGestureHandlerRootView (Android) to make sure that nesting it inside RNGestureHandlerEnabledRootView doesn't crash it, which can happen if the user already configured their Mainactivity.java.
Test plan

Tested by wrapping one of the examples in the example app in GestureHandlerRootView and whole app in gestureHandlerRootHOC, both with and without the setup code in MainActivity.java

Questions

Should we remove the MainActivity.java step from the docs and only use the component/hoc? Arguably it's easier to setup than MainActivity.java because you don't have to touch native code.