fix visibility issue when starting drag
Was caused because view was hidden before the drag events began to propagate causing a situation that  if a drag has started but no events would be emitted because the finger stopped touching the screen the view would remain invisible.
This was fixed by:
Moving `setVisibility(false)` to `onHandle` instead of `startDragging`.
Reinstating `onHandle(MotionEvent event)` to the super pan handler so it can handle this exact case of `ACTION_UP` fired after `startDragging` and propagate state changes