Copy Directions enum to JS (#1366)
A common problem when importing RNGH is not finding native modules when running on not-yet-supported platforms like React Native for iOS or React Native for Windows. 

Moving Directions into JS allows using it directly in tests. Also, State (which is defined in the Native Module, too) was already copied to JS.

Please note that importing mappings from native modules doesn't incur Bridge roundtrips, this is merely for tests and other platforms' convenience.