Fix compatibility with React 18 (#1500)
On web, there are cases now in React 18 where componentWillUnmount gets called twice in a row. I think it is related to reactwg/react-18#31. This causes a No handler for tag error since the handler was cleaned up already in the first componentWillUnmount call.

To workaround the issue I added a check to no-op of the handler tag is not found. I think this is fine in that case and simpler than tracking whether the handler was cleaned up in the createHandler component.