Update docs (#1604)
Adds basic docs about the new api. Parts about gestures are nearly identical to those about gesture handlers with slight naming modifications and type annotations. The docs will need to be updated when we solidify naming convention and update types.
Moves minPointers prop only to Pan and Tap gestures and gesture handlers as they are the only ones that use it.
Changes type in FlingGesture direction to be consistent with the docs.