Add collapsable prop to the child of the Wrap (#1636)
Modifies the Wrap component used by the GestureDetector to find the viewTag of the view it should attach to. By adding the collapsable: false prop it makes sure that the child view of the wrap ends up in the view hierarchy.
Note that there are still problems when there is a View used just to group components (no props) inside a functional component that is a child of a GestureDetector.