Fix LongPressGestureHandler inconsistency (#1610)
Update LongPressGestureHandler on iOS to now include BEGAN in its state flow.
Fix long press duration calculation (mach_timebase_info returns time in device-specific units)

Previously on iOS the maxDist prop only applied before handler activation. When the handler was active, the pointer could be moved further than maxDist without cancelling, while handler and this was not the case on Android. This PR makes it follow the behavior of the Android one, however we might want to consider adding a prop that would enable this behavior on both platforms. (Or even make it a default behavior).