Update PanGestureHandler on iOS (#1611)
Update PanGestureHandler on iOS to send UNDETERMINED -> BEGAN state change event on pointer down, not just before activation.
Use self.view.window instead of self.view as an argument for [self translationInView] in order to use absolute position to calculate translation instead of the relative one. This fixes a bug where if an Animated.View was a direct child of a PanGestureHandler and it was rotated and/or scaled, the transform would influence the pan gesture (i.e. the axis of the gesture would be rotated with the component, or translation would be scaled with respect to the scale of the transformation).