Fix gesture handlers not working when children overflow their parent (#1618)
On Android to get all handlers that could want events from the pointer the view tree is traversed and when a view contains the pointer, all handlers attached to that view are extracted. This works fine as long as children do not overflow the parent, when that happens the handlers attached to the ancestors may not be extracted because the pointer may not be inside them (but may be inside a child view).
This change adds a check inside the method responsible for extracting handlers to see if the current view is overflowing its parent. It it is, all gesture handlers attached to its ancestor views are also extracted (it may be nested deeper than one level).

Fixes https://github.com/software-mansion/react-native-gesture-handler/issues/1532