Remove animatedGesture prop, use animated gesture automatically (#1665)
Remove animatedGesture prop from GestureDetector component and add automatic check whether the gesture callbacks are worklets. If any one of them is a worklet, gesture will be initialized the same way it would be via animatedGesture. If no callback is a worklet, then the gesture will be initialized without using Reanimated.