Get rid of specifying optionalDependencies in package.json
The optionalDependencies option was causing reanimated to be installed in other projects as indirect dependency because optionals need to be specified at the project scope and not at the dependency scope. As a result projects using pre-released versions would end up with two versions of reanimated installed under node_modules which in turn was causing a number of different issues specifically around reanimated's ConfigHelper functionality.