Prepare state management for pointer events (#1670)
Add common interface and protocol that will be shared between Gesture Handler and Reanimated.
Add setGestureState method to allow managing state of the gesture handlers. This method will be used by react-native-reanimated to allow for synchronous state management inside worklets.
Extracted pre- and post-activation logic to separate functions on Android. This allows for it to be executed just by calling activate.
Reset Pinch and Rotation state on activation to prevent unexpected behavior with manual activation.
Add manual activation implementation on Android and iOS