Don't cancel handler if enabled didn't change (#1722)
On every re-render the updated handler config is send to the native side. On Android, if the enabled prop was present in the config, this causes the handler to get cancelled even when enabled was "changed" from true to true. This PR changes it so that the handler gets cancelled only when the value actually changes.