Change invokation order (#1751)
Changed order of method invocations in `reset` method of gesture recognizers. `triggerAction` should be called before resetting the `pointerTracker`, as it in turn may reset the GestureHandler they are associated with. This causes GestureHandler to set its cached `lastState` to `UNDETERMINED` which was breaking the state flow when using `touchEvents`.