Add onFinalize callback (#1749)
Change onEnd callback to be called only when the gesture finishes in the ACTIVE state.
Add onFinalize callback to all gestures. It's called whenever the gestures finishes, if the gesture transitions from ACTIVE state, it will be called after onEnd.

This allows for simpler logic in callbacks, as onEnd can rely on logic in onStart, and onFinalize can rely on logic in onBegin.