Don't send UNDETERMINED -> CANCELLED event (#1796)
This PR changes Orchestrator in a way that the `UNDETERMINED` -> `CANCELLED` state change event would not be sent. Before introduction of touch events such change couldn't occur, but now it's possible to activate the gesture at the moment the finger touches the screen. Events are delivered to handlers sequentially, so it may happen that handlers next in line would get cancelled before receiving the first event (thus before they can initialize). If this doesn't result in a crash, then the `UNDETERMINED` to `CANCELLED` state change event will be sent, triggering `onFinalize` callback without calling `onBegin` first.