Fix awaiting handler not ending (#1774)
On Android when a handler awaiting another one tries to activate its state is changed to ACTIVE but event with that change is delayed until the other finishes. If the awaiting handler were to fail before that happens, state change event with information about its failure would not be sent.

This PR adds a workaround that sends state change event when the handler is active internally but still awaiting another one.