Send first `onTouchesDown` after gesture begins (#1793)
On Android touch events are sent before other events. This causes problems when the gesture is activated in the onTouchesDown callback. The gesture then goes from UNDETERMINED state to ACTIVE before it has a change to process the first event. For example, if done with PanGesture, the translation would start with values equal to the coordinates of the pointer.

This PR adds a workaround that sends the first touch event (which would be TOUCH_DOWN) only after the handler has processed the event (making sure the handler is initialized correctly, and UNDETERMINED -> BEGAN state change event is sent).