Fix race condition between the orchestrator and the registry (#1684)
The handlers collection must have been modified in other place and the registry is the only class that modifies it. My guess is that the handler was dropped or removed from view and relevant method was called (asynchronously)
causing the list to be modified while it was being iterated over. This PR locks the list while it's being modified or iterated over.