Rebuild Reanimated event only when needed (#1856)
At the moment, Reanimated event is rebuilt on every update which may cause a heavy performance hit in some cases. I believe that it's a leftover from pre-release Gesture Handler 2. This PR changes it so that the event will be rebuilt only when the gestures are reattached alleviating performance issues.

Should fix #1849.