Fix ripple animation on touchables in some cases (#1877)
In cases where `TouchableOpacity` and `TouchableNativeFeedback` were used together, but `TouchableOpacity` was placed first, the color of ripple animation would also be set to `transparent` for `TouchableNativeFeedback`. This PR changes the behavior of GestureHandlerButton so that it will set the color of ripple animation to [`android.R.attr.colorControlHighlight`](https://developer.android.com/reference/android/R.attr#colorControlHighlight) when no color is set by the user.