Fix overriding `testID` when `createHandler` props do not have one assigned (#1910)
Fix related to comment [here](https://github.com/software-mansion/react-native-gesture-handler/pull/1844#issuecomment-1057400208) about `RNGestureHandlerButton` with `handlerType="NativeViewGestureHandler"` having missing testIDs when they were previously provided and assigned. The `testID` on the child being cloned within `createHandler` was being removed by the `testID` being set within the render function even the `testID` on the handler was not set. Now it will only assign a `testID` as a prop on the cloned element if a `testID` was provided to the handler.