Rename `sendStateChangeEvent` to `sendEvent` (#1939)
Before 0d1a917 there was a method responsible for sending each type of event, however `sendStateChangeEvent` was used to send all of them. In 0d1a917 the redundant methods were removed, but the name remained.
This PR changes the name of `sendStateChangeEvent` to `sendEvent` on iOS with hope that it will be (slightly) less confusing.