Effectively remove `RNGestureHandlerEnabledRootView` (#1936)
`tryInitializeHandlerForReactRootView` was causing crashes, especially with combination with CodePush. `RNGestureHandlerEnabledRootView` was deprecated in version 2.0.0 in favor of `GestureHandlerRootView` component and this PR removes implementation of `RNGestureHandlerEnabledRootView` and code responsible for interactions with it.
Adds a new page in docs explaining how to migrate to `GestureHandlerRootView`.

Potentially fixes https://github.com/software-mansion/react-native-gesture-handler/issues/1675.