Fix gestures not attaching to views on iOS when using Fabric (#1979)
When gestures are created and attached to views, on iOS the operations are batched and executed on the UI thread. The method we are using currently is run too early - the batched queue is flushed before the gestures are created on the JS side, effectively making the app unresponsive.

This PR adds a method that allows flushing the queue from the JS side, after all the operations are scheduled.