Fix nested buttons with outer one disabled on iOS (#1991)
Fixes https://github.com/software-mansion/react-native-gesture-handler/issues/294.

When the outer button had `enabled` property set to `false` it was mapped to the `enabled` property of `UIControl` with the same name. In that case the view would ignore all touches on its surface. This PR changes that by:
- mapping the `enabled` prop to the new `userEnabled` property of `RNGestureHandlerButton`
- changing the logic of hit-testing to check for the `userEnabled` prop and handle it correctly
Now the `hitTest` method finds a deepest view that was under the pointer, and searches the view tree upwards until a view that can handle the event is found.

This change makes it consistent with Android.