Use constant value for translation when `DrawerLayout` is idle (#2005)
Fixes https://github.com/software-mansion/react-native-gesture-handler/issues/386.

`AnimatedInterpolation` was causing problems on Android when the activity was paused for some reason (like permission or share dialog). For some reason it was changing its value on pause, opening the drawer, and again on resume closing the drawer. This PR changes `DrawerLayout` so that it uses constant value instead of interpolation when the drawer is idle.