Don't update shared value unless the config has changed (#2037)
This PR slightly modifies how gesture callbacks are updated when using Reanimated. Instead of updating them every time, it will now be updated only when the gesture config has changed. In combination with `useMemo`, this will make optimizing performance-critical use cases easier.

It accomplishes this by adding `gestureId` property to every gesture instance, which then is checked before updating the shared value. The update happens only when the id is different (which will happen only when the gesture inside `useMemo` is recalculated, i.e. only when the dependencies change).