Update  scroll view class check in `RNNativeViewHandler` on Fabric (#2031)
When `NativeViewGestureHandler` is attached to a `ScrollView` on iOS, it sets its `delaysContentTouches` property to `YES`. This is done by checking if the view the gesture is attached to is `RCTScrollView` class, which wraps `UIScrollView`, but on Fabric the `UIScrollView` is wrapped with `RCTScrollViewComponentView`. This PR makes it so that the former is used on Paper and the later on Fabric.