Fix `simultaneousHandlers` prop on `NativeViewGestureHandler` on iOS (#2018)
It seems like to make two `UIGestureRecognizers` work simultaneously, both calls of `shouldRecognizeSimultaneouslyWithGestureRecognizer` should return `YES` (should `A` be recognized alongside `B` and vice versa). This isn't much of a problem in case of all gestures but `NativeViewGestureHandler`, which most often is used implicitly via components exported by Gesture Handler. In this case one can access `simultaneousHandlers` prop on the wrapped component, but cannot obtain the ref to the `NativeViewGestureHandler` (only to the view).

This PR makes `shouldRecognizeSimultaneouslyWithGestureRecognizer` symmetric, so only one gesture needs to be marked as simultaneous with the other